refactored migrations into independent classes
