+++ protocols/migration/__init__.py
@@ -5,6 +5,12 @@ from .base_migration import BaseMigration
 from .migration_participant_1_0_0_to_participant_1_0_3 import MigrationParticipants100To103
 from .migration_participant_1_0_3_to_participant_1_1_0 import MigrationParticipants103To110
 from .migration_reports_3_0_0_to_participant_1_0_0 import MigrationReports3ToParticipant1
+from .migration_participant_1_0_1_to_reports_3_0_0 import MigrationParticipants101ToReports
+from .migration_participant_1_0_3_to_participant_1_0_0 import MigrationParticipants103To100
+from .migration_participant_1_0_0_to_reports_3_0_0 import MigrationParticipants100ToReports
+from .migration_participant_1_1_0_to_participant_1_0_0 import MigrateParticipant110To100
+from .migration_participant_1_1_0_to_participant_1_0_3 import MigrateParticipant110To103
+from .migration_reports_4_0_0_to_reports_3_0_0 import MigrateReports400To300
 
 from .migration_reports_3_0_0_to_reports_4_0_0 import MigrateReports3To4
 from .migration_reports_4_0_0_to_reports_5_0_0 import MigrateReports400To500
