+++ protocols/tests/test_migration/test_migration_reports_5_0_0_to_reports_4_0_0.py
@@ -8,10 +8,10 @@ from protocols import participant_1_1_0
 from protocols.util.dependency_manager import VERSION_61
 from protocols.util.factories.avro_factory import FactoryAvro
 from protocols.util.factories.avro_factory import GenericFactoryAvro
-from protocols.migration.participants import MigrationParticipants103To100
+from protocols.migration import MigrationParticipants103To100
 from protocols.tests.test_migration.base_test_migration import TestCaseMigration
-from protocols.migration.migration_reports_5_0_0_to_reports_4_0_0 import MigrateReports500To400
-from protocols.migration.migration_participant_1_1_0_to_participant_1_0_3 import MigrateParticipant110To103
+from protocols.migration import MigrateReports500To400
+from protocols.migration import MigrateParticipant110To103
 
 
 class TestMigrateReports5To400(TestCaseMigration):
